CHAT_CAPSTONE/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.10
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langgraph==0.2.35
langgraph-checkpoint==2.0.1
langgraph-checkpoint-postgres==2.0.1
langgraph-checkpoint-sqlite==2.0.0
langsmith==0.1.134
marshmallow==3.22.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.51.2
orjson==3.10.7
packaging==24.1
propcache==0.2.0
psycopg==3.2.3
psycopg2-binary
psycopg-pool==3.2.3
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
tenacity==8.5.0
tiktoken==0.8.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
yarl==1.15.2
